job_application_bot/requirements.txt
requests>=2.31.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
pandas>=2.2.0
lxml>=4.9.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0
tenacity>=8.2.0
pydantic>=2.5.0
click>=8.1.0
colorama>=0.4.6
openpyxl>=3.1.0
flask>=3.0.0
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.0
flask-wtf>=1.2.0
wtforms>=3.1.0
flask-migrate>=4.0.0
